Cast all quantities to single precision by default (#472)
* suppress all default MDSplus casting to use native precision

* remove several castings to double

* remove safe_cast routine and mds cast operations

* cast everything by default into float32 and int32

* cast sql data down to float32

* fix cast when reading floats from csv

* Revert "Remove dipprog_dt_rt from DIII-D XFAIL column (#399)"

This reverts commit 6d0cce98faeed48efdfdb8a7c972ef7ecbc26df4.

* cast times to float32 well ahead of xr conversion

* compare only to 0.1 ms absolute precision

* Update d3d test_shared_time_setting ground truth

* remove h_alpha from DIII-D xfailures

---------

+++ disruption_py/core/physics_method/params.py
@@ -64,7 +64,7 @@ class PhysicsMethodParams:
         return to_tuple(
             data={
                 "shot": len(self.times) * [self.shot_id],
-                "time": self.times,
+                "time": self.times.astype("float32"),
             },
             dim="idx",
         )
